refactor(tui): extract format/style helpers to app/format.rs (S4 PR 4)
Fourth slice. The 24 free-fn formatting helpers (token-usage rendering,
context-pressure summarization, the LineTone-keyed styled/plain/title
family, JSON shape helpers) move into app/format.rs (~196 LoC). Pure
functions; no AppState access. 84/84 TUI tests pass.

Mechanics:
  - new file: crates/memphis-tui/src/app/format.rs (196 LoC)
  - app/mod.rs: 2530 → 2359 LoC (-171)
  - 24 functions marked pub(super); re-imported into app/mod.rs's
    namespace via `use format::{...}` so existing call sites in
    mod.rs and sibling submodules continue to resolve unchanged
  - Two test-only helpers (format_status_pressure,
    format_status_token_usage) gated behind #[cfg(test)] use
  - mod.rs imports cleaned up: removed now-unused
    TelegramReadinessSummary (used only by format.rs) and serde_json
    Value alias (only json! macro is used in mod.rs)

LoC trajectory across S4:
  app.rs (start)       6039
  → after PR 1         5395  (-644 → commands.rs)
  → after PR 2         4100  (-1295 → host_results.rs)
  → after PR 3         2530  (-1570 → tests.rs)
  → after PR 4         2359  (-171 → format.rs)

ADR-001 ceiling for crates/ is 2000 LoC. One more slice (likely the
render_* family + telegram_*_lines into app/render.rs) clears it.
